Created a separate package.json for td.server, moved folder structure and updated relative file references
