+++ td.server/.jshintignore
@@ -1,0 +1,1 @@
+test/
