+++ td.server/.jshintrc
@@ -1,0 +1,3 @@
+{
+    "node": true
+}
