+++ td.server/test/config/parsers.config_spec.js
@@ -12,7 +12,7 @@ describe('request parser tests', function() {
     beforeEach(function() {
         
         app = express();
-        require('../../../td.server/config/parsers.config')(app);
+        require('../../src/config/parsers.config')(app);
         
     });
