+++ td.server/test/support/jasmine.json
@@ -1,5 +1,5 @@
 {
-  "spec_dir": "./td.tests/serverspec/",
+  "spec_dir": "./test/",
   "spec_files": [
     "**/*[sS]pec.js"
   ]
